LibG: Fix small things

+++ libs/libg/Rect.cpp
@@ -2,7 +2,7 @@
 
 namespace LG {
 
-Rect::Rect(int x, int y, int width, int height)
+Rect::Rect(int x, int y, size_t width, size_t height)
     : m_x(x)
     , m_y(y)
     , m_width(width)
